requirements.txt
PyQt5==5.15.10
kafka-python
protobuf
grpcio-tools
